Update firehosex result attempted, successful, failed to int64

+++ awsex/service/firehosex/result.go
@@ -2,7 +2,7 @@ package firehosex
 
 // Result firehose result struct
 type Result struct {
-	Attempted  int `json:"attempted"`
-	Successful int `json:"successful"`
-	Failed     int `json:"failed"`
+	Attempted  int64 `json:"attempted"`
+	Successful int64 `json:"successful"`
+	Failed     int64 `json:"failed"`
 }
